requirements.txt
# Runtime
fastapi==0.136.0
uvicorn==0.44.0
pydantic==2.13.2
requests==2.33.1
apscheduler==3.11.2
reportlab==4.4.10

# Config
python-dotenv==1.1.0

# Testing
pytest==8.3.5
pytest-mock==3.14.0
